requirements.txt
llama-cpp-python[server]
openai